use plotfix on data, to fill in missing points
it's pretty unlikely that we'll use zero power, and since
that implies an infinite pulse interval, it's even less
likely that we can measure it.  so we use it to signify
"missing data", and tell gnuplot to try and plot 1/0 when
it sees it -- i.e., an illegal value.  this causes a gap
in the plot.