don't bother doing averaging to create the watts_now value.
just use the interval between the current and previous pulses.
previously i was thinking there was error in the timing, and
there isn't, really.  read_ir timestamps the pulses, so there's
some error there, but not much, and it's predominantly a constant
error, the same for every pulse.  (thanks to n17ikh on hackaday
for pointing this out.)